requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-simple-history>=3.4
psycopg2-binary>=2.9
python-dateutil>=2.8
pandas>=2.0
python-dotenv>=1.0
